✨ feat: UnclassifiedViewModel → ItemListViewModel 재사용 컴포넌트로 교체 #84
